requirements.txt
alembic==1.12.1
attrs==24.2.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
flasgger==0.9.7.1
Flask==2.2.5
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
mistune==3.0.2
mypy==0.942
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
packaging==24.0
parameterized==0.8.1
pkgutil_resolve_name==1.3.10
protobuf==3.19.4
pycodestyle==2.5.0
pycparser==2.21
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.1
PyYAML==6.0.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.2.3
Flask-JWT-Extended
zipp==3.7.0
